Đẩy món ăn lên firebase

+++ app/src/main/java/com/project/pro112/hydrateam/thepolycoffee/activity/shopping/Purchase.java
@@ -20,6 +20,8 @@ import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.MarkerOptions;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
 import com.project.pro112.hydrateam.thepolycoffee.R;
 import com.project.pro112.hydrateam.thepolycoffee.models.OrderedFood;
 import com.project.pro112.hydrateam.thepolycoffee.tempdatabase.tempdatabase;
@@ -28,6 +30,8 @@ import java.text.DateFormat;
 import java.text.DecimalFormat;
 import java.util.ArrayList;
 import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
 
 import static com.project.pro112.hydrateam.thepolycoffee.activity.shopping.Order.linearButtonViewCart;
 
@@ -36,6 +40,8 @@ public class Purchase extends AppCompatActivity implements View.OnClickListener,
     private Button order;
     private GoogleMap mMap;
     private TextView total;
+    private boolean isPushDataDone;
+    private Double totalp;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -47,9 +53,9 @@ public class Purchase extends AppCompatActivity implements View.OnClickListener,
         setUpHideButtonWhenSrollToTheBottom();
         setTotal();
         setUpMaps();
-        Toast.makeText(this, "User id: "+getUserId()+"", Toast.LENGTH_SHORT).show();
-        String currentDateTimeString = DateFormat.getDateTimeInstance().format(new Date());
-        Toast.makeText(this, "Date: "+currentDateTimeString, Toast.LENGTH_SHORT).show();
+//        Toast.makeText(this, "User id: "+getUserId()+"", Toast.LENGTH_SHORT).show();
+//        String currentDateTimeString = DateFormat.getDateTimeInstance().format(new Date());
+//        Toast.makeText(this, "Date: "+currentDateTimeString, Toast.LENGTH_SHORT).show();
     }
 
     private String getUserId() {
@@ -71,12 +77,13 @@ public class Purchase extends AppCompatActivity implements View.OnClickListener,
         order.setOnClickListener(this);
         scrollView = (ScrollView) findViewById(R.id.scrollView);
         total = (TextView) findViewById(R.id.total);
+        isPushDataDone = false;
     }
 
     public void setTotal() {
         LinearLayout linearLayout = linearButtonViewCart;
         TextView tv = (TextView) linearLayout.getChildAt(0);
-        Double totalp = Double.parseDouble(tv.getText().toString().substring(0, tv.getText().toString().length() - 1));
+        totalp = Double.parseDouble(tv.getText().toString().substring(0, tv.getText().toString().length() - 1));
         if(totalp > 200000){
 
         }else{
@@ -135,12 +142,13 @@ public class Purchase extends AppCompatActivity implements View.OnClickListener,
     @Override
     public void onClick(View v) {
         tempdatabase tempdatabase = new tempdatabase(getBaseContext());
-        ArrayList<OrderedFood> orderedFoods = tempdatabase.getOrderedFoods();
-        Toast.makeText(this, "nut purchase kich hoat....Delete all trash data", Toast.LENGTH_SHORT).show();
-        for (int i = 0; i < orderedFoods.size(); i++) {
-            Toast.makeText(this, "Day len fire base: "+orderedFoods.get(i).getName(), Toast.LENGTH_SHORT).show();
+        PushDataToFireBase();
+        if(isPushDataDone) {
+            tempdatabase.deleteAlldata();
+            Toast.makeText(this, "Mua hàng hành công!", Toast.LENGTH_SHORT).show();
+        }else{
+            Toast.makeText(this, "Đã xảy ra lỗi, mua hàng thất bại!", Toast.LENGTH_SHORT).show();
         }
-        tempdatabase.deleteAlldata();
     }
 
     //Set up maps
@@ -157,5 +165,43 @@ public class Purchase extends AppCompatActivity implements View.OnClickListener,
         LatLng sydney = new LatLng(-34, 151);
         mMap.addMarker(new MarkerOptions().position(sydney).title("Marker in Sydney"));
         mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(sydney,15));
+    }
+
+    private void PushDataToFireBase() {
+        tempdatabase tempdatabase = new tempdatabase(getBaseContext());
+        ArrayList<OrderedFood> orderedFoods = tempdatabase.getOrderedFoods();
+        if(getUserId() !=null && orderedFoods.size() > 0) {
+            // get firebase instance
+            FirebaseDatabase database = FirebaseDatabase.getInstance();
+            DatabaseReference myRef;
+            myRef = database.getReference();
+            myRef.child("Orders").child(""+getUserId());
+            Map myMapFoods = new HashMap();
+            if(orderedFoods.size() >= 0) {
+                for (int i = 0; i < orderedFoods.size(); i++) {
+                    OrderedFood food = new OrderedFood(orderedFoods.get(i).getDiscription()+"",
+                            ""+orderedFoods.get(i).getImage(),
+                            ""+orderedFoods.get(i).getName(),
+                            orderedFoods.get(i).getPrice(),
+                            orderedFoods.get(i).getAmount());
+                    myMapFoods.put("Food"+(i+1), food);
+                }
+            }else{
+                Toast.makeText(this, "Vui lòng chọn món!", Toast.LENGTH_SHORT).show();
+            }
+
+            Map mParent = new HashMap();
+            mParent.put("Date", ""+getCurrentTime());
+            mParent.put("Total", totalp);
+            mParent.put("Foods", myMapFoods);
+            myRef.child("Orders").child(""+getUserId()).push().setValue(mParent);
+        }else{
+            Toast.makeText(this, "Vui lòng chọn ít nhất 1 món!", Toast.LENGTH_SHORT).show();
+        }
+        isPushDataDone = true;
+    }
+    private String getCurrentTime() {
+        String currentDateTimeString = DateFormat.getDateTimeInstance().format(new Date());
+        return currentDateTimeString;
     }
 }
